gazou-syoukei/requirements.txt
opencv-python
numpy
pillow
openai